refactor(headers): move public headers to include/lance/
Convention is namespaced subdir (<lance/lance.h>) to avoid
header-name collisions when installed system-wide. Matches
librsvg's <librsvg/rsvg.h> layout.

Breaking change for external consumers vendoring lance.h
from a git checkout.
